Implement customer menus, fix app and logger problems

+++ src/main/java/com/revature/Entities/AccountDao.java
@@ -16,7 +16,7 @@ public class AccountDao implements Dao<Account> {
     public void insert(Account account) {
         try {
             PreparedStatement pStatement = connection.prepareStatement("insert into accounts values (nextval('account_seq'), ?, ?, ?, ?");
-            pStatement.setInt(1, account.getBalance());
+            pStatement.setDouble(1, account.getBalance());
             pStatement.setInt(2, account.getAccountType());
             pStatement.setBoolean(3, account.isApproved());
             pStatement.setInt(4, account.getPermission());
@@ -36,7 +36,7 @@ public class AccountDao implements Dao<Account> {
             ResultSet resultSet = statement.executeQuery("select * from accounts");
             while (resultSet.next()) {
                 account = new Account();
-                account.setBalance(resultSet.getInt("balance"));
+                account.setBalance(resultSet.getDouble("balance"));
                 account.setAccountType(resultSet.getInt("accounttype"));
                 account.setApproved(resultSet.getBoolean("approved"));
                 account.setAccountNumber(resultSet.getInt("accountnumber"));
@@ -58,7 +58,7 @@ public class AccountDao implements Dao<Account> {
             ResultSet resultSet = statement.executeQuery("select * from accounts where username=" + account.accountOwner);
             while (resultSet.next()){
                 account = new Account();
-                account.setBalance(resultSet.getInt("balance"));
+                account.setBalance(resultSet.getDouble("balance"));
                 account.setAccountType(resultSet.getInt("accounttype"));
                 account.setApproved(resultSet.getBoolean("approved"));
                 account.setAccountNumber(resultSet.getInt("accountnumber"));
@@ -80,7 +80,7 @@ public class AccountDao implements Dao<Account> {
     public void update(Account account) {
         try {
             PreparedStatement pStatement = connection.prepareStatement("update accounts set balance=? where accountowner=? and accounttype=?");
-            pStatement.setInt(1, account.getBalance());
+            pStatement.setDouble(1, account.getBalance());
             pStatement.setString(2, account.getAccountOwner());
             pStatement.setInt(3, account.getAccountType());
             pStatement.executeUpdate();
